refactor: reorganize environment files and update configuration values for PPO
