+++ Dockerfile.train
@@ -7,11 +7,9 @@ WORKDIR /app
 
 COPY train.py ./
 COPY env_factory.py ./
-COPY worker_env.py ./
-COPY manager_env.py ./
 COPY custom_env.py ./
-COPY curriculum_env.py ./
-COPY dr_spcrl.py ./
+COPY hrl ./hrl
+COPY dr_spcrl ./dr_spcrl
 COPY lattice ./lattice
 
 COPY config.py ./
